refactor: moving reset code to component folder
